add explicit file system registration
